+++ CHANGELOG.md
@@ -1,4 +1,3 @@
 * Add idle animation to player models
 * Add shivering to player models when frozen
-* Use cached player skins
-* Fix player models going dark
+* Fix Ping Wheel and Xaero's Minimap mixin crash